% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacon_short.R
\name{bacon_short}
\alias{bacon_short}
\title{Bacon}
\usage{
bacon_short(x)
}
\arguments{
\item{bmiq_results_file}{}
}
\value{
A list with lamdas for before and after and tibble with bacon added
}
\description{
Bacon
}
