% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_parameters_cross.R
\name{best_parameters_cross}
\alias{best_parameters_cross}
\title{Performs cross validation and returns the best parameters values of SVM}
\usage{
best_parameters_cross(x, y, v, w)
}
\arguments{
\item{x}{- classifications data}

\item{y}{- decision values}

\item{v}{- number of folds}

\item{w}{- weight}
}
\value{
The best values of gamma and C
}
\description{
Performs cross validation and returns the best parameters values of SVM
}
