% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_explicit_mortality2}
\alias{calibrate_explicit_mortality2}
\title{Calibrates explicit mortality by Amin}
\usage{
calibrate_explicit_mortality2(n_sim = 10^7)
}
\arguments{
\item{n_sim}{number of agents}
}
\value{
difference in mortality rates and life table
}
\description{
Calibrates explicit mortality by Amin
}
