% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_mi_incidence}
\alias{calibrate_mi_incidence}
\title{Calibrates MI incidence.}
\usage{
calibrate_mi_incidence(simple = F)
}
\arguments{
\item{simple}{true or false.}
}
\value{
TODO not implemented yet.
}
\description{
Calibrates MI incidence.
}
