% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Cget_output}
\alias{Cget_output}
\title{Main outputs of the current run.}
\usage{
Cget_output()
}
\value{
number of agents, cumulative time, number of deaths, number of COPD cases, as well as exacerbation statistics and QALYs.
}
\description{
Main outputs of the current run.
}
