% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_improvement.R
\name{check_improvement}
\alias{check_improvement}
\title{This function checks if there is an improvement based on HV metric;}
\usage{
check_improvement(previous.front, current.front)
}
\arguments{
\item{previous.P}{- nondominated solutions in the objective space of the solution set obtained 5 evaluations ago}

\item{current.P}{- nondominated solutions in the objective space of the current solution set}
}
\value{
TRUE or FALSE
}
\description{
This function checks if there is an improvement based on HV metric;
}

