% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_summaries.R
\name{col_summaries}
\alias{col_summaries}
\title{Summaries for multiple columns in a df}
\usage{
col_summaries(df)
}
\arguments{
\item{df}{}
}
\value{
A data frame with mean, min, max and range for each column of a df
}
\description{
Summaries for multiple columns in a df
}
