% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_init_simplex.R
\name{create_init_simplex}
\alias{create_init_simplex}
\title{Generates a simplex from the evaluated decicion vectors and coresponding scalarized objective value.}
\usage{
create_init_simplex(Xe, Ye, Ys, P)
}
\arguments{
\item{Xe}{- an k x N matrix of evaluated vectors in the decison space, where N is a number of decision variables}

\item{Ye}{- an k x M matrix of evaluated vectors in the objective space, where M is a number of objective functions}

\item{Ys}{- a vector of lenght k containing scalarized objective values}

\item{P}{- a list returned by function pareto_filter}
}
\value{
initial simplex used in a local search; it's a matrix (N+1) x (N+1), where rows correspond to vertices, first N columns are decision variables, last column - scalarized function value
}
\description{
Generates a simplex from the evaluated decicion vectors and coresponding scalarized objective value.
}

