% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTLZ.R
\name{dtlz1}
\alias{dtlz1}
\title{Set of 9 DTLZ test functions for the case D -> D-9}
\usage{
dtlz1(x)
}
\arguments{
\item{x}{is a vector of length n such that 0 < x_i < 1, i=1,...,n}
}
\value{
a vector of length M = n - 9
}
\description{
Set of 9 DTLZ test functions for the case D -> D-9
}

