% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicR.R
\docType{package}
\name{epicR}
\alias{epicR}
\title{\code{epicR} package}
\description{
Evaluation Platform in COPD (EPIC) R Package
}
\details{
See the README on
\href{https://github.com/aminadibi/epicR#readme}{GitHub}
}
