% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{export_figures}
\alias{export_figures}
\title{Runs the model and exports an excel file with all output data}
\usage{
export_figures(nPatients = 10^4)
}
\arguments{
\item{nPaients}{number of agents}
}
\value{
an excel file with all output
}
\description{
Runs the model and exports an excel file with all output data
}
