% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_run}
\alias{export_run}
\title{Runs the model and exports an excel file with all output data}
\usage{
export_run(nPatients = 10^4)
}
\arguments{
\item{nPaients}{number of agents}
}
\value{
an excel file with all output
}
\description{
Runs the model and exports an excel file with all output data
}
