% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{express_matrix}
\alias{express_matrix}
\title{Express matrix.}
\usage{
express_matrix(mtx)
}
\arguments{
\item{mtx}{a matrix}
}
\value{
No return value, called for side effects (prints R code to console).
}
\description{
Takes a named matrix and writes the R code to populate it;
useful for generating input expressions from calibration results.
}
