% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDMRcpgs.R
\name{extractDMRcpgs}
\alias{extractDMRcpgs}
\title{Extract individual dmCpGs from your regression DMR object}
\usage{
extractDMRcpgs(dmr_object, dmr_number)
}
\arguments{
\item{dmr_object}{The DMR object from the regression}

\item{dmr_number}{DMR number x in the table}
}
\value{
A table of the CpGs within the DMR
}
\description{
Extract individual dmCpGs from your regression DMR object
}
