% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_centroid.R
\name{find_centroid}
\alias{find_centroid}
\title{Finds the centroid of the simplex}
\usage{
find_centroid(iter)
}
\arguments{
\item{iter}{- variable containing all the information related with the current EPIC iteration}
}
\value{
iter - updated information of the the current EPIC iteration
}
\description{
Finds the centroid of the simplex
}

