% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_report.R
\name{gen_report}
\alias{gen_report}
\title{Generating a report - takes a limm atable file only}
\usage{
gen_report(output_file_name, raw_results_path)
}
\arguments{
\item{raw_results_path}{}
}
\value{
Rmarkdown report in an outs folder in reports folder
}
\description{
Generating a report - takes a limm atable file only
}
