% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_weigths.R
\name{generate_weights}
\alias{generate_weights}
\title{Generates normalized weight vector of k dimensions and s divisions, so that each weight vector is near
the previous one. It uses the method of generating reflected k-ary Gray codes.}
\usage{
generate_weights(s, k)
}
\arguments{
\item{s}{- divisions}

\item{k}{- dimension}
}
\value{
The normalized weight vectors
}
\description{
Generates normalized weight vector of k dimensions and s divisions, so that each weight vector is near
the previous one. It uses the method of generating reflected k-ary Gray codes.
}
\examples{
generate_weights(10, 2)
generate_weights(5, 3)
}

