% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/core.R
\name{get_agent_events}
\alias{get_agent_events}
\title{Returns all events of an agent.}
\usage{
get_agent_events(id)

get_agent_events(id)
}
\arguments{
\item{id}{Agent number}
}
\value{
all events of agent \code{id}

dataframe consisting all events specific to agent \code{id}
}
\description{
Returns all events of an agent.

Returns events specific to an agent.
}
