% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/core.R
\name{get_agent_size_bytes}
\alias{get_agent_size_bytes}
\title{Returns the size of agent struct in bytes}
\usage{
get_agent_size_bytes()

get_agent_size_bytes()
}
\value{
size of agent struct in bytes

size in bytes
}
\description{
Returns the size of agent struct in bytes

Get size of agent struct in bytes (from C code)
}
