% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/core.R
\name{get_all_events}
\alias{get_all_events}
\title{Returns all events.}
\usage{
get_all_events()

get_all_events()
}
\value{
all events

dataframe consisting all events.
}
\description{
Returns all events.

Returns all events.
}
