% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_all_events_matrix}
\alias{get_all_events_matrix}
\title{Returns a matrix containing all events}
\usage{
get_all_events_matrix()
}
\value{
a matrix containing all events
}
\description{
Returns a matrix containing all events
}
