% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/core.R
\name{get_events_by_type}
\alias{get_events_by_type}
\title{Returns all events of a certain type.}
\usage{
get_events_by_type(event_type)

get_events_by_type(event_type)
}
\arguments{
\item{event_type}{event_type number}
}
\value{
all events of the type \code{event_type}

dataframe consisting all events of the type \code{event_type}
}
\description{
Returns all events of a certain type.

Returns certain events by type
}
