% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{get_input}
\alias{get_input}
\title{Returns a list of default model input values}
\usage{
get_input(
  age0 = 40,
  time_horizon = 20,
  discount_cost = 0,
  discount_qaly = 0.03,
  closed_cohort = 0,
  jurisdiction = "canada"
)
}
\arguments{
\item{age0}{Starting age in the model}

\item{time_horizon}{Model time horizon}

\item{discount_cost}{Discounting for cost outcomes}

\item{discount_qaly}{Discounting for QALY outcomes}

\item{closed_cohort}{Whether the model should run as closed_cohort, open-population by default.}

\item{jurisdiction}{Jurisdiction for model parameters ("canada" or "us")}
}
\value{
A list with four components:
  \describe{
    \item{values}{A nested list of model input parameter values}
    \item{help}{A nested list of help text descriptions for each parameter}
    \item{ref}{A nested list of reference information for each parameter}
    \item{config}{The raw configuration object loaded from the JSON file}
  }
}
\description{
Returns a list of default model input values
}
