% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_progress}
\alias{get_progress}
\title{Returns current simulation progress.}
\usage{
get_progress()
}
\value{
Number of agents processed so far (last_id).
}
\description{
Returns current simulation progress.
}
