% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_runtime_stats}
\alias{get_runtime_stats}
\title{Returns run time stats.}
\usage{
get_runtime_stats()
}
\value{
agent size as well as memory and random variable fill stats.
}
\description{
Returns run time stats.
}
