% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_sample_output}
\alias{get_sample_output}
\title{Returns a sample output for a given year and gender.}
\usage{
get_sample_output(year, sex)
}
\arguments{
\item{year}{a number}

\item{sex}{a number, 0 for male and 1 for female}
}
\value{
that specific output
}
\description{
Returns a sample output for a given year and gender.
}
