% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-management.R
\name{get_user_config_dir}
\alias{get_user_config_dir}
\title{Get user config directory path}
\usage{
get_user_config_dir()
}
\value{
Path to user config directory
}
\description{
Returns the path to the user configuration directory following
CRAN policy using \code{tools::R_user_dir()}.
}
