% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{getExResults}
\alias{getExResults}
\title{Plot Exacerbations}
\usage{
getExResults()
}
\arguments{
\item{nPatients}{number of patients/agents/individuals to simulate}

\item{arg1}{string: one of "sex", "year", "gold"}

\item{arg2}{string: one of "sex", "year", "gold"}

\item{exacType}{string: "all" or "severe"}

\item{perCapita}{numeric: if null, parameter is ignored, otherwise, population size to look at}
}
\value{
Returns a plot
}
\description{
Creates plots of exacerbations
}
