% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_feasible.R
\name{is_feasible}
\alias{is_feasible}
\title{Checks if a vector V is feasible in the decision space bounded by box constraints}
\usage{
is_feasible(V, L, U)
}
\arguments{
\item{V}{- a decision vector}

\item{L}{- lower bound, a vector}

\item{U}{- upper bound, a vector}
}
\value{
True, if vector V is feasible, FALSE, otherwise
}
\description{
Checks if a vector V is feasible in the decision space bounded by box constraints
}
\examples{
is_feasible(c(3,5), c(2,1), c(5,4))
is_feasible(c(3,5), c(2,1), c(5,6))
}

