% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_feasible_con.R
\name{is_feasible_con}
\alias{is_feasible_con}
\title{Checks if a vector V is feasible in the decision space constrained by function con}
\usage{
is_feasible_con(con, V)
}
\arguments{
\item{con}{- constraints function ( g <= 0)}

\item{V}{- a decision vector}
}
\value{
True, if vector V is feasible; FALSE, otherwise
}
\description{
Checks if a vector V is feasible in the decision space constrained by function con
}

