% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-incidence.R
\name{iterate_COPD_inc}
\alias{iterate_COPD_inc}
\title{Solves stochastically for COPD incidence rate equation.}
\usage{
iterate_COPD_inc(nIterations = 1000, nPatients = 10000, time_horizon = 20)
}
\arguments{
\item{time_horizon}{in years}

\item{n_Iterations}{number of iterations for the numberical solution}

\item{n_Patients}{number of simulated agents.}
}
\value{
regression co-efficients as files
}
\description{
Solves stochastically for COPD incidence rate equation.
}
