% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_EPIC_NM.R
\name{iterate_EPIC_NM}
\alias{iterate_EPIC_NM}
\title{Performs one iteration of EPIC algorithm}
\usage{
iterate_EPIC_NM(wv, i, con, L, U, p.next, p.pareto, strategy, iter, absmin,
  absmax, local.search)
}
\arguments{
\item{wv}{- a weighting vector}

\item{i}{- iteration number}

\item{con}{- constraints, an analytical function cheap to evaluate, in a form g_i(x)>=0, if available; otherwise it is equal to FALSE}

\item{L}{- row vector of lower bounds of the design space}

\item{U}{- row vector of upper bounds of the design space}

\item{p.next}{- probability used to select a next decision vector}

\item{p.pareto}{- indicator used to assign points to one of the set: dominated or nondominated}

\item{strategy}{- strategy used to select the nex vector to evaluate}

\item{iter}{- a list conatinig information related with the current algorithm iteration}

\item{absmin}{- the estimate of minimum values of objective functions (if not provided we will use the min value of Y)}

\item{absmax}{- the estimate of maximum values of objective functions (if not provided we will use the max value of Y)}

\item{local.search}{- indicates what EPIC version will be run. TRUE - enhanced EPIC version,FALSE -  the original one. DEFAULT = TRUE.}
}
\value{
iter -  an updated list
}
\description{
Performs one iteration of EPIC algorithm
}
