% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-management.R
\name{list_available_jurisdictions}
\alias{list_available_jurisdictions}
\title{List available config jurisdictions}
\usage{
list_available_jurisdictions()
}
\value{
Character vector of available jurisdictions
}
\description{
List available config jurisdictions
}
