% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_colnames.R
\name{list_colnames}
\alias{list_colnames}
\title{List all the colnames in alphabetical order as a tibble}
\usage{
list_colnames(df)
}
\arguments{
\item{df}{}
}
\value{
printed tibble in console of colnames
}
\description{
List all the colnames in alphabetical order as a tibble
}
