% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_sig_cps_plot.R
\name{mm_plot_0.25}
\alias{mm_plot_0.25}
\title{Plot results from missMethyl analysis of significant array hits}
\usage{
mm_plot_0.25(x)
}
\arguments{
\item{x}{}
}
\value{
A table of missMrthyl results
}
\description{
Plot results from missMethyl analysis of significant array hits
}
\examples{
mm_plot_0.25(sig_0.25)
}
