% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm_sig_cps.R
\name{mm_sig_cpgs}
\alias{mm_sig_cpgs}
\title{Perform missMethyl on array hits with an FDR <0.5}
\usage{
mm_sig_cpgs(sig_hits)
}
\arguments{
\item{sig_hits}{}
}
\value{
A data frame with the top 50 most significant results from all GO ontologies.
}
\description{
Perform missMethyl on array hits with an FDR <0.5
}
