% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnormArma}
\alias{mvrnormArma}
\title{Samples from a multivariate normal}
\usage{
mvrnormArma(n, mu, sigma)
}
\arguments{
\item{n}{number of samples to be taken}

\item{mu}{the mean}

\item{sigma}{the covariance matrix}
}
\value{
the multivariate normal sample
}
\description{
Samples from a multivariate normal
}
