% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_graphs.R
\name{normality_graphs}
\alias{normality_graphs}
\title{Show normality graphs for a given variable}
\usage{
normality_graphs(df, variable)
}
\arguments{
\item{df}{A data frame}

\item{variable}{A continuous variable}
}
\value{
Plot with 4 graphs highlighting normality
}
\description{
Show normality graphs for a given variable
}
