% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-management.R
\name{open_user_config_dir}
\alias{open_user_config_dir}
\title{Open user config directory in file explorer}
\usage{
open_user_config_dir()
}
\description{
Open user config directory in file explorer
}
