% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_significance.R
\name{overlap_significance}
\alias{overlap_significance}
\title{Determines significance of overlap using hypergeometric probability test}
\usage{
overlap_significance(
  table_a,
  table_b,
  follow_bacon = FALSE,
  top_100 = NULL,
  fdr = NULL
)
}
\arguments{
\item{fdr}{}
}
\value{
GeneOverlap result to console
}
\description{
Determines significance of overlap using hypergeometric probability test
}
