% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_filter.R
\name{pareto_filter}
\alias{pareto_filter}
\title{Generates normalized weight vector of k dimensions and s divisions, so that each weight vector is near
the previous one. It uses the method of generating reflected k-ary Gray codes.}
\usage{
pareto_filter(y, x)
}
\arguments{
\item{y}{- objective values corresponding to \code{x}}

\item{x}{- decision values}
}
\value{
yp - Pareto front in the objective space, xp - Pareto set in the decision space,
ip - logical indicator, TRUE if solution is nondominated, FALSE if solution is dominated
}
\description{
Generates normalized weight vector of k dimensions and s divisions, so that each weight vector is near
the previous one. It uses the method of generating reflected k-ary Gray codes.
}

