% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{reset_model_input}
\alias{reset_model_input}
\title{Reset the cached model input}
\usage{
reset_model_input()
}
\description{
Forces the default model_input to be reloaded on next access.
Useful after modifying config files.
}
