% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{resume}
\alias{resume}
\title{Resumes running of model.}
\usage{
resume(max_n_agents = NULL)
}
\arguments{
\item{max_n_agents}{maximum number of agents}
}
\value{
0 if successful.
}
\description{
Resumes running of model.
}
