% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{sanity_check}
\alias{sanity_check}
\title{Basic tests of model functionalty. Serious issues if the test does not pass.}
\usage{
sanity_check()
}
\value{
tests results
}
\description{
Basic tests of model functionalty. Serious issues if the test does not pass.
}
