% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner.R
\name{see_txt_corner}
\alias{see_txt_corner}
\title{See corner of big text file}
\usage{
see_txt_corner(file_path)
}
\arguments{
\item{file_path}{}
}
\value{
Prints corner of big text file in console
}
\description{
See corner of big text file
}
