% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_strategy.R
\name{selection_strategy}
\alias{selection_strategy}
\title{Selects a decision vector from a set representing the decision space}
\usage{
selection_strategy(s, Xu, Xe, prob.non, p.next, iter, m)
}
\arguments{
\item{s}{- strategy ID, from 1 to 4}

\item{Xu}{- unevaluated decision vectors}

\item{Xe}{- evaluated decision vectors}

\item{prob.non}{- probabilities of unevaluated vectors to be nondominated}

\item{p.next}{- probability of nondominance the candidate vector should have to be selected to evaluate next}

\item{iter}{- iteration number}

\item{m}{- number of vectors to be selected selected for evaluation}
}
\value{
I - index of the selectec decision vector from Xu
}
\description{
Selects a decision vector from a set representing the decision space
}
