% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_input_var}
\alias{set_input_var}
\title{Sets input variables.}
\usage{
set_input_var(name, value)
}
\arguments{
\item{name}{a string}

\item{value}{a number}
}
\value{
0 if successful
}
\description{
Sets input variables.
}
