% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_settings_var}
\alias{set_settings_var}
\title{Sets model settings.}
\usage{
set_settings_var(name, value)
}
\arguments{
\item{name}{a name}

\item{value}{a value}
}
\value{
0 if successful.
}
\description{
Sets model settings.
}
