% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrink_simplex.R
\name{shrink_simplex}
\alias{shrink_simplex}
\title{This function performs simplex shrink operation. It replaces all but the best vertice by newly calculated}
\usage{
shrink_simplex(iter, delta, con, L, U)
}
\arguments{
\item{iter}{- a list conatinig information related with the current algorithm iteration}

\item{delta}{- a constant used to shrink the simplex edges}

\item{con}{- constraints, an analytical function cheap to evaluate, in a form g_i(x)>=0, if available; otherwise it is equal to FALSE}

\item{L, U}{- row vectors of lower and upper bounds of the design space}
}
\value{
iter -  an updated list
}
\description{
This function performs simplex shrink operation. It replaces all but the best vertice by newly calculated
}

