% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tcheby_values.R
\name{Tcheby_values}
\alias{Tcheby_values}
\title{Converts a multiobjective problem into a scalarized single objective using weighted augmented Chebyshev approach}
\usage{
Tcheby_values(y, gwv, absmin, absmax)
}
\arguments{
\item{y}{- objective values of a multiobjective problem}

\item{gwv}{- a weigthing vector}

\item{absmin}{- minimal values of objectives}

\item{absmax}{- maximal values of objectives}
}
\value{
k - objective value of a scalarized problem
}
\description{
Converts a multiobjective problem into a scalarized single objective using weighted augmented Chebyshev approach
}

