% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{terminate_session}
\alias{terminate_session}
\title{Terminates a session and releases allocated memory.}
\usage{
terminate_session()
}
\value{
0 if successful.
}
\description{
Terminates a session and releases allocated memory.
}
