% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{test_case_detection}
\alias{test_case_detection}
\title{Returns results of Case Detection strategies}
\usage{
test_case_detection(n_sim = 10000, p_of_CD = 0.1, min_age = 40,
  min_pack_years = 0, only_smokers = 0, OR_of_CD = 1.82)
}
\arguments{
\item{n_sim}{number of agents}

\item{p_of_CD}{probability of recieving case detection given that an agent meets the selection criteria}

\item{min_age}{minimum age that can recieve case detection}

\item{min_pack_years}{minimum pack years that can recieve case detection}

\item{only_smokers}{set to 1 if only smokers should recieve case detection}

\item{OR_of_CD}{odds ratio for the impact of recieving case detection on the odds of being diagnosed}
}
\value{
results of case detection strategy compared to no case detection
}
\description{
Returns results of Case Detection strategies
}
