% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_simplex.R
\name{transform_simplex}
\alias{transform_simplex}
\title{Updates the simplex with obtained new values and proceed with the following simplex operation untill
the new decision vector is generated and requires a function evaluation}
\usage{
transform_simplex(iter, L, U, con)
}
\arguments{
\item{iter}{- a list conatining all the information related with the current algorithm iteration}

\item{L, }{U - row vectors of lower and upper bounds of the design space}

\item{con}{- constraints, an analytical function cheap to evaluate, in a form g_i(x)>=0, if available; otherwise it is equal to FALSE}
}
\value{
an updated list with the information about the current EPIC algorithm iteration
}
\description{
Updates the simplex with obtained new values and proceed with the following simplex operation untill
the new decision vector is generated and requires a function evaluation
}

