% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universe.R
\name{universe}
\alias{universe}
\title{Get universe og genes without duplicates}
\usage{
universe(var)
}
\arguments{
\item{bmiq_doc}{}
}
\value{
A character string of around 23k entries
}
\description{
Get universe og genes without duplicates
}
