% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_simplex.R
\name{update_simplex}
\alias{update_simplex}
\title{Updates the simplex used in a local search of EPIC algorithm}
\usage{
update_simplex(iter, x, y)
}
\arguments{
\item{iter}{- list containing all the information about the current EPIC algorithm iteration}

\item{x}{- simplex vertice in the decision space}

\item{y}{- scalarized response in the objective space corresponding to the \code{x}}
}
\value{
an updated list with the information about the current EPIC algorithm iteration
}
\description{
Updates the simplex used in a local search of EPIC algorithm
}
