% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_comorbidity}
\alias{validate_comorbidity}
\title{Returns results of validation tests for comorbidities}
\usage{
validate_comorbidity(n_sim = 1e+05)
}
\arguments{
\item{n_sim}{number of agents}
}
\value{
validation test results
}
\description{
Returns results of validation tests for comorbidities
}
