% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-management.R
\name{validate_config}
\alias{validate_config}
\title{Validate a config file}
\usage{
validate_config(jurisdiction = "canada", user = TRUE)
}
\arguments{
\item{jurisdiction}{Jurisdiction to validate}

\item{user}{Whether to validate user config (TRUE) or package config (FALSE)}
}
\value{
TRUE if valid, FALSE otherwise (with warnings about issues)
}
\description{
Validate a config file
}
