% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_COPDUS}
\alias{validate_COPDUS}
\title{Returns results of validation tests for US COPD Prevalence}
\usage{
validate_COPDUS()
}
\value{
validation test results for the US
}
\description{
Returns results of validation tests for US COPD Prevalence
}
