% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_diagnosis}
\alias{validate_diagnosis}
\title{Returns results of validation tests for diagnosis}
\usage{
validate_diagnosis(n_sim = 10000, jurisdiction = "canada")
}
\arguments{
\item{n_sim}{number of agents}

\item{jurisdiction}{character string specifying the jurisdiction ("canada" or "us"). Default is "canada"}
}
\value{
validation test results
}
\description{
This function returns a table showing the proportion of COPD patients diagnosed
over the model's runtime. It also provides a second table that breaks down the proportion
of diagnosed patients by COPD severity. Additionally, the function generates a plot
to visualize the distribution of diagnoses over time.
}
