% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_exacerbation_by_sex}
\alias{validate_exacerbation_by_sex}
\title{Validates exacerbation trends by sex over time}
\usage{
validate_exacerbation_by_sex(n_sim = 10000)
}
\arguments{
\item{n_sim}{number of simulated agents (default=1e4)}
}
\value{
a plot showing exacerbation counts by sex over time
}
\description{
Validates exacerbation trends by sex over time
}
