% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_gpvisits}
\alias{validate_gpvisits}
\title{Returns results of validation tests for GP visits}
\usage{
validate_gpvisits(n_sim = 10000, jurisdiction = "canada")
}
\arguments{
\item{n_sim}{number of agents}
}
\value{
validation test results
}
\description{
This function returns Average number of GP visits by sex, COPD severity and
COPD diagnosis status along with their plots.
}
