% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_lung_function}
\alias{validate_lung_function}
\title{Returns results of validation tests for lung function}
\usage{
validate_lung_function()
}
\value{
validation test results
}
\description{
Returns results of validation tests for lung function
}
