% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_medication}
\alias{validate_medication}
\title{Returns results of validation tests for medication module.}
\usage{
validate_medication(n_sim = 50000)
}
\arguments{
\item{n_sim}{number of agents}
}
\value{
validation test results for medication
}
\description{
Returns results of validation tests for medication module.
}
