% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_overdiagnosis}
\alias{validate_overdiagnosis}
\title{Returns results of validation tests for overdiagnosis}
\usage{
validate_overdiagnosis(n_sim = 10000, jurisdiction = "canada")
}
\arguments{
\item{n_sim}{number of agents}

\item{jurisdiction}{character string specifying the jurisdiction ("canada" or "us"). Default is "canada"}
}
\value{
validation test results
}
\description{
This function returns the proportion of non-COPD subjects overdiagnosed
over model time.
}
