% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_populationUS}
\alias{validate_populationUS}
\title{Returns plots comparing EPIC population projection vs US population data}
\usage{
validate_populationUS()
}
\value{
plots showing population projections vs actual values for 40-59, 60-79, 80+ age groups
}
\description{
This function validates the population simulation by comparing the EPIC model population
outputs with US population data over a set time horizon. It generate 3 plots for
the following age groups : 40-59, 60-79, 80+
}
