% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_symptoms}
\alias{validate_symptoms}
\title{Returns results of validation tests for Symptoms}
\usage{
validate_symptoms(n_sim = 10000, jurisdiction = "canada")
}
\arguments{
\item{n_sim}{number of agents}

\item{jurisdiction}{character string specifying the jurisdiction ("canada" or "us"). Default is "canada"}
}
\value{
validation test results
}
\description{
This function plots the prevalence of cough, dyspnea, phlegm and wheeze
over time and by GOLD stage.
}
