% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whataremydmrs.R
\name{whataremydmrs}
\alias{whataremydmrs}
\title{Extract DMRs from the regression DMR object}
\usage{
whataremydmrs(dmr_object)
}
\arguments{
\item{dmr_object}{}
}
\value{
Returns a df with DMRs arranged by stouffers value
}
\description{
Extract DMRs from the regression DMR object
}
