% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicR.R
\docType{package}
\name{epicR-package}
\alias{epicR-package}
\alias{epicR}
\title{\code{epicR} package}
\description{
Evaluation Platform in COPD (EPIC) R Package
}
\details{
See the README on
\href{https://github.com/aminadibi/epicR#readme}{GitHub}
}
\author{
\strong{Maintainer}: Amin Adibi \email{adibi@alumni.ubc.ca}

Authors:
\itemize{
  \item Mohsen Sadatsafavi \email{mohsen.sadatsafavi@ubc.ca} [copyright holder]
  \item Kate Johnson \email{kate.johnson@alumni.ubc.ca}
}

}
