% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_agent_size_bytes}
\alias{get_agent_size_bytes}
\title{Get size of agent struct in bytes (from C code)}
\usage{
get_agent_size_bytes()
}
\value{
size in bytes
}
\description{
Get size of agent struct in bytes (from C code)
}
