% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{get_list_elements}
\alias{get_list_elements}
\title{Get list elements}
\usage{
get_list_elements(ls, running_name = "")
}
\arguments{
\item{ls}{ls}

\item{running_name}{running_name}
}
\description{
Get list elements
}
