% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-management.R
\name{get_user_config_dir}
\alias{get_user_config_dir}
\title{Get user config directory path}
\usage{
get_user_config_dir()
}
\value{
Path to user config directory
}
\description{
Get user config directory path
}
