% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{init_session}
\alias{init_session}
\title{Initializes a model. Allocates memory to the C engine.}
\usage{
init_session(settings = get_default_settings())
}
\arguments{
\item{settings}{customized settings.}
}
\value{
0 if successful.
}
\description{
Initializes a model. Allocates memory to the C engine.
}
