% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{run}
\alias{run}
\title{Runs the model, after a session has been initialized.}
\usage{
run(max_n_agents = NULL, input = NULL)
}
\arguments{
\item{max_n_agents}{maximum number of agents}

\item{input}{customized input criteria}
}
\value{
0 if successful.
}
\description{
Runs the model, after a session has been initialized.
}
