% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_COPD}
\alias{validate_COPD}
\title{Returns results of validation tests for COPD}
\usage{
validate_COPD(incident_COPD_k = 1, return_CI = FALSE)
}
\arguments{
\item{incident_COPD_k}{a number (default=1) by which the incidence rate of COPD will be multiplied.}

\item{return_CI}{if TRUE, returns 95 percent confidence intervals for the "Year" coefficient}
}
\value{
validation test results
}
\description{
Returns results of validation tests for COPD
}
