% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_COPDUS}
\alias{validate_COPDUS}
\title{Returns results of validation tests for US COPD Prevalence}
\usage{
validate_COPDUS()
}
\arguments{
\item{incident_COPD_k}{a number (default=1) by which the incidence rate of COPD will be multiplied.}

\item{return_CI}{if TRUE, returns 95 percent confidence intervals for the "Year" coefficient}
}
\value{
validation test results for the US
}
\description{
Returns results of validation tests for US COPD Prevalence
}
