% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_diagnosis}
\alias{validate_diagnosis}
\title{Returns results of validation tests for diagnosis}
\usage{
validate_diagnosis(n_sim = 10000)
}
\arguments{
\item{n_sim}{number of agents}
}
\value{
validation test results
}
\description{
Returns results of validation tests for diagnosis
}
