% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_smokingUS}
\alias{validate_smokingUS}
\title{Returns results of validation tests for smoking module in the US population.}
\usage{
validate_smokingUS(remove_COPD = 1, intercept_k = NULL)
}
\arguments{
\item{remove_COPD}{0 or 1. whether to remove COPD-related mortality.}

\item{intercept_k}{a number}
}
\value{
validation test results
}
\description{
Returns results of validation tests for smoking module in the US population.
}
