% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_problems.R
\name{zdt1}
\alias{zdt1}
\title{Set of multiobjective benchamrk test including ZDT test functions, OKA2, Viennet and TF4 (a constrained problem)}
\usage{
zdt1(x)
}
\arguments{
\item{x}{- is a decision vector of lenght \code{n}}
}
\value{
function output as a vector of lenght 2 or 3
}
\description{
Set of multiobjective benchamrk test including ZDT test functions, OKA2, Viennet and TF4 (a constrained problem)
}
\examples{
oka2(c(1,2,3))
zdt1(c(2,2,2))
}

